completed react accordion component
